--- projects/bamm/abi.json
@@ -1,4 +1,0 @@
-{
-    "bammsArray": "address[]:bammsArray",
-    "getTVL": "function getTVL(address) view returns (address,address,uint256,uint256)"
-}
